requirements.txt
flask==2.2.3
python-dotenv==1.0.0
requests==2.28.2
openai==0.27.0
gunicorn==20.1.0
werkzeug==2.2.3 